+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v2.2.11
 
 * SQL: Change max primary key in DDL to 767 to fix MySQL
+* Elasticsearch: use an inline Groovy script to efficiently delete a property; also fixes query bugs
 
 # v2.2.10